v1 of porting over to precompute -- removed a lot of tests for guarded deploy that need to be added back in at some point

+++ src/UnifapV2/libraries/UnifapV2Library.sol
@@ -49,7 +49,7 @@ library UnifapV2Library {
         returns (uint256[] memory amounts)
     {
         require(path.length >= 2, "UniswapV2Library: INVALID_PATH");
-        amounts = new uint[](path.length);
+        amounts = new uint256[](path.length);
         amounts[0] = amountIn;
         for (uint256 i; i < path.length - 1; i++) {
             (uint256 reserveIn, uint256 reserveOut) = getReserves(factory, path[i], path[i + 1]);
@@ -64,7 +64,7 @@ library UnifapV2Library {
         returns (uint256[] memory amounts)
     {
         require(path.length >= 2, "UniswapV2Library: INVALID_PATH");
-        amounts = new uint[](path.length);
+        amounts = new uint256[](path.length);
         amounts[amounts.length - 1] = amountOut;
         for (uint256 i = path.length - 1; i > 0; i--) {
             (uint256 reserveIn, uint256 reserveOut) = getReserves(factory, path[i - 1], path[i]);
